♻️[refactor]: 파일 이동 및 삼품 수정 페이지 이미지 로직 수정
