+++ OpenMarket/OpenMarket/Controller/enum/CurrencyType.swift
