+++ OpenMarket/OpenMarket/Controller/enum/LayoutType.swift
