+++ OpenMarket/OpenMarket/Model/Product.swift
@@ -13,11 +13,19 @@ struct Product: Decodable {
     let thumbnail: String?
     let currency: String?
     let price: Int?
+    let description: String?
     let bargainPrice: Int?
     let discountedPrice: Int?
     let stock: Int?
     let createdAt: String?
     let issuedAt: String?
-    let images: [Image]?
-    let descriptions: String?
+    let images: [Images]?
+}
+
+struct Images: Codable {
+    let id: Int?
+    let url: String?
+    let thumbnailUrl: String?
+    let succeed: Bool?
+    let issuedAt: String?
 }
